backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "playwright>=1.57.0",
    "playwright-stealth>=2.0.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
